Feat: 스케줄 선택 페이지 404페이지 추가
